complete average grade mapred job

+++ hadoop/average_grade/src/main/java/org/mdp/learn/hadoop/average_grade/AverageGradeReducer.java
@@ -17,7 +17,7 @@ public class AverageGradeReducer extends Reducer<CourseAndStudentWritable, IntWr
       count++;
     }
 
-    context.write(key, new FloatWritable(sum / count));
+    context.write(key, new FloatWritable((float) sum / count));
   }
 
 }
